+++ app/models/quote.rb
@@ -13,4 +13,14 @@ class Quote < ApplicationRecord
            dependent:  :destroy
 
   has_many :coupons, class_name: 'Bonds::Coupon', inverse_of: :quote, dependent: :destroy
+
+  def coupons_sum_for_time_range(time_range=1.year)
+    beginning_of_day = DateTime.now.beginning_of_day
+    Rails.cache.fetch("quotes/#{id}/coupons_sum/#{beginning_of_day}") do
+      coupons
+        .where('payment_date > ? AND payment_date < ?', beginning_of_day, beginning_of_day + time_range)
+        .pluck(:coupon_value)
+        .sum
+    end
+  end
 end